# docs/swagger.yaml
basePath: /
definitions:
...
  entities.MealPlan:
    properties:
      backup_amount:
        type: integer
      backup_menu_id:
        type: string
      created_at:
        type: string
      is_allergy:
        type: boolean
      main_amount:
        type: integer
      meal_plan_id:
        type: string
      meal_type:
        description: breakfast, lunch, dinner
        type: string
      menu_id:
        type: string
    type: object
  entities.Menu:
    properties:
      description:
        description: such as ingredients
        type: string
      menu_id:
        type: string
      menu_name:
        type: string
    type: object
...
  models.CreateMealPlanRequest:
    properties:
      backup_amount:
        type: integer
      backup_menu_id:
        type: string
      human_in_the_loop:
        type: boolean
      main_amount:
        type: integer
      meal_type:
        type: string
      menu_id:
        type: string
    required:
    - main_amount
    - meal_type
    - menu_id
    type: object
  models.CreateMenuRequest:
    properties:
      description:
        type: string
      menu_name:
        type: string
    required:
    - description
    - menu_name
    type: object
...
info:
  contact: {}
  description: This is a sample for New-Acis API.
  title: New-Acis API
paths:
...
  /api/meals/meal-plans:
    get:
      consumes:
      - application/json
      description: Get all meal plans. Only users with Kitchen Staff, Super User,
        or Admin role can manage meals.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              result:
                items:
                  $ref: '#/definitions/entities.MealPlan'
                type: array
              status:
                type: string
              status_code:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Get All Meal Plans
      tags:
      - Meal
    post:
      consumes:
      - application/json
      description: Create a new meal plan. Only users with Kitchen Staff, Super User,
        or Admin role can manage meals.
      parameters:
      - description: Meal plan information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateMealPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              message:
                type: string
              result:
                $ref: '#/definitions/entities.MealPlan'
              status:
                type: string
              status_code:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Create Meal Plan
      tags:
      - Meal
  /api/meals/meal-plans/{id}:
    get:
      consumes:
      - application/json
      description: Get a meal plan by ID. Only users with Kitchen Staff, Super User,
        or Admin role can manage meals.
      parameters:
      - description: Meal Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              result:
                $ref: '#/definitions/entities.MealPlan'
              status:
                type: string
              status_code:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Get Meal Plan by ID
      tags:
      - Meal
  /api/meals/meal-plans/manual:
    post:
      consumes:
      - application/json
      description: Create a new meal plan in manual mode (AI allergy check is skipped).
        Only users with Kitchen Staff, Super User, or Admin role can manage meals.
      parameters:
      - description: Meal plan information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateMealPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              message:
                type: string
              result:
                $ref: '#/definitions/entities.MealPlan'
              status:
                type: string
              status_code:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Create Meal Plan Manual
      tags:
      - Meal
  /api/meals/meal-plans/today:
    get:
      consumes:
      - application/json
      description: Get all meal plans created today. Only users with Kitchen Staff,
        Super User, or Admin role can manage meals.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              result:
                items:
                  $ref: '#/definitions/entities.MealPlan'
                type: array
              status:
                type: string
              status_code:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Get Today's Meal Plans
      tags:
      - Meal
  /api/meals/menus:
    get:
      consumes:
      - application/json
      description: Get all menus. Only users with Kitchen Staff, Super User, or Admin
        role can manage meals.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              result:
                items:
                  $ref: '#/definitions/entities.Menu'
                type: array
              status:
                type: string
              status_code:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Get All Menus
      tags:
      - Meal
    post:
      consumes:
      - application/json
      description: Create a new menu. Only users with Kitchen Staff, Super User, or
        Admin role can manage meals.
      parameters:
      - description: Menu information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              message:
                type: string
              result:
                $ref: '#/definitions/entities.Menu'
              status:
                type: string
              status_code:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              result: {}
              status:
                type: string
              status_code:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Create Menu
      tags:
      - Meal
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"